requirements.txt
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
jupyter==1.0.0
notebook==6.5.4
ipython==8.11.0
polars==0.19.0
psycopg2-binary==2.9.6
pyarrow==12.0.1
scipy==1.10.1